+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- `errors.new_class` options checking
+
 ## [2.1.0] - 2018-05-27
 
 ### Fixed